KG-14303: Picked up dependency on 5.0 components. Updated license headers in all the source files. Changed the package name and the imports. Added NOTICE.txt, LICENSE.txt, and README.md files.
