+++ NOTICE.txt
@@ -1,0 +1,10 @@
+This product depends on Kaazing WebSocket Gateway - AMQP Java Client 5.0
+
+	License:	http://www.apache.org/licenses/LICENSE-2.0 (Apache License, Version 2.0)
+	Homepage:	https://github.com/kaazing/amqp.client.java.git
+
+This product depends on Kaazing WebSocket Gateway - Java Client 5.0
+
+	License:	http://www.apache.org/licenses/LICENSE-2.0 (Apache License, Version 2.0)
+	Homepage:	https://github.com/kaazing/gateway.client.java.git
+
