+++ pom.xml
@@ -1,23 +1,44 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <!--
-
-    Copyright (c) 2007-2012, Kaazing Corporation. All rights reserved.
-
+** This is free and unencumbered software released into the public domain.
+**
+** Anyone is free to copy, modify, publish, use, compile, sell, or
+** distribute this software, either in source code form or as a compiled
+** binary, for any purpose, commercial or non-commercial, and by any
+** means.
+**
+** In jurisdictions that recognize copyright laws, the author or authors
+** of this software dedicate any and all copyright interest in the
+** software to the public domain. We make this dedication for the benefit
+** of the public at large and to the detriment of our heirs and
+** successors. We intend this dedication to be an overt act of
+** relinquishment in perpetuity of all present and future rights to this
+** software under copyright law.
+**
+** THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND,
+** EXPRESS OR IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF
+** MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT.
+** IN NO EVENT SHALL THE AUTHORS BE LIABLE FOR ANY CLAIM, DAMAGES OR
+** OTHER LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE,
+** ARISING FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR
+** OTHER DEALINGS IN THE SOFTWARE.
+**
+** For more information, please refer to <http://unlicense.org/>
 -->
 <project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
     <modelVersion>4.0.0</modelVersion>
 
     <parent>
-        <groupId>com.kaazing.gateway.core</groupId>
-        <artifactId>com.kaazing.gateway.core.common</artifactId>
-        <version>1.0.0.47</version>
+        <groupId>org.kaazing</groupId>
+        <artifactId>gateway.client.java.common</artifactId>
+        <version>5.1.0.5</version>
     </parent>
 
-    <groupId>com.kaazing.gateway.amqp</groupId>
-    <artifactId>com.kaazing.gateway.amqp.client.java.demo</artifactId>
-    <name>Kaazing WebSocket Gateway - AMQP Edition Java Client Demo</name>
+    <groupId>org.kaazing</groupId>
+    <artifactId>amqp.client.java.demo</artifactId>
+    <name>Kaazing WebSocket Gateway - AMQP Java Client Demo</name>
     <packaging>jar</packaging>
-    <version>4.0.0.12-SNAPSHOT</version>
+    <version>5.0.0.1-SNAPSHOT</version>
 
     <url>https://github.com/kaazing/amqp.client.java.demo.git</url>
     <scm>
@@ -28,30 +49,24 @@
 
     <dependencies>
         <dependency>
-            <groupId>com.kaazing.gateway.amqp</groupId>
-            <artifactId>com.kaazing.gateway.amqp.client.java</artifactId>
+            <groupId>org.kaazing</groupId>
+            <artifactId>amqp.client.java</artifactId>
             <type>jar</type>
-            <version>[4.0.0.0,4.1.0.0)</version>
+            <version>[5.0.0.0,5.1.0.0)</version>
         </dependency>
          <dependency>
-            <groupId>com.kaazing.gateway</groupId>
-            <artifactId>com.kaazing.gateway.client.java</artifactId>
+            <groupId>org.kaazing</groupId>
+            <artifactId>gateway.client.java</artifactId>
             <type>jar</type>
-            <version>[4.0.0.0,4.1.0.0)</version>
+            <version>[5.0.0.0,5.1.0.0)</version>
          </dependency>
-        <dependency>
+        <!--dependency>
             <groupId>com.kaazing.gateway.core</groupId>
             <artifactId>com.kaazing.gateway.core.demo.resources</artifactId>
             <version>[4.0.0.0,4.1.0.0)</version>
             <type>zip</type>
             <classifier>headers</classifier>
-        </dependency>
-        <dependency>
-            <groupId>com.kaazing</groupId>
-            <artifactId>tools.code.signing</artifactId>
-            <version>[4.0.0.0,4.1.0.0)</version>
-            <type>jar</type>
-        </dependency>
+        </dependency-->
     </dependencies>
 
     <build>
@@ -72,7 +87,7 @@
                             <shadedClassifierName>shaded</shadedClassifierName>
                             <transformers>
                                 <transformer implementation="org.apache.maven.plugins.shade.resource.ManifestResourceTransformer">
-                                    <mainClass>com.kaazing.net.ws.amqp.demo.AmqpFrame</mainClass>
+                                    <mainClass>org.kaazing.net.ws.amqp.demo.AmqpFrame</mainClass>
                                 </transformer>
                             </transformers>
                             <filters>
@@ -107,55 +122,6 @@
                    </archive>
                 </configuration>
             </plugin>
-            <plugin>
-                <groupId>org.apache.maven.plugins</groupId>
-                <artifactId>maven-dependency-plugin</artifactId>
-                <executions>
-                    <execution>
-                        <id>unpack-code-signing</id>
-                        <phase>process-sources</phase>
-                        <goals>
-                            <goal>unpack-dependencies</goal>
-                        </goals>
-                        <configuration>
-                            <includeGroupIds>com.kaazing</includeGroupIds>
-                            <includeArtifactIds>tools.code.signing</includeArtifactIds>
-                            <includeTypes>jar</includeTypes>
-                            <outputDirectory>target/codesigning</outputDirectory>
-                        </configuration>
-                    </execution>
-                </executions>
-            </plugin>
-            <plugin>
-                <groupId>org.apache.maven.plugins</groupId>
-                <artifactId>maven-jarsigner-plugin</artifactId>
-                <version>1.3.2</version>
-                <executions>
-                    <execution>       
-                        <phase>package</phase>
-                        <id>sign</id>
-                        <goals>
-                            <goal>sign</goal>
-                        </goals>
-                    </execution>
-                    <execution>
-                        <id>verify</id>
-                        <goals>
-                            <goal>verify</goal>
-                        </goals>
-                    </execution>
-                </executions>
-                <configuration>
-                    <keystore>target/codesigning/kaazing-code-signing-cert.pfx</keystore>
-                    <alias>1</alias>
-                    <storetype>pkcs12</storetype>
-                    <storepass>kaazing</storepass>
-                    <arguments>
-                        <argument>-tsa</argument>
-                        <argument>https://timestamp.geotrust.com/tsa</argument>
-                    </arguments>
-                </configuration>
-            </plugin> 
 
             <!-- Copy the demo resources -->
             <plugin>
@@ -201,6 +167,14 @@
                         </configuration>
                     </execution>
                 </executions>
+            </plugin>
+            <plugin>
+                <groupId>com.google.code.maven-license-plugin</groupId>
+                <artifactId>maven-license-plugin</artifactId>
+                <configuration>
+                    <useDefaultExcludes>true</useDefaultExcludes>
+                    <failIfMissing>false</failIfMissing>
+                </configuration>
             </plugin>
         </plugins>
     </build>
