+++ src/main/java/org/kaazing/net/ws/amqp/demo/AmqpPanel.java
@@ -1,8 +1,31 @@
 /**
- * Copyright (c) 2007-2014, Kaazing Corporation. All rights reserved.
- */
+** This is free and unencumbered software released into the public domain.
+**
+** Anyone is free to copy, modify, publish, use, compile, sell, or
+** distribute this software, either in source code form or as a compiled
+** binary, for any purpose, commercial or non-commercial, and by any
+** means.
+**
+** In jurisdictions that recognize copyright laws, the author or authors
+** of this software dedicate any and all copyright interest in the
+** software to the public domain. We make this dedication for the benefit
+** of the public at large and to the detriment of our heirs and
+** successors. We intend this dedication to be an overt act of
+** relinquishment in perpetuity of all present and future rights to this
+** software under copyright law.
+**
+** THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND,
+** EXPRESS OR IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF
+** MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT.
+** IN NO EVENT SHALL THE AUTHORS BE LIABLE FOR ANY CLAIM, DAMAGES OR
+** OTHER LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE,
+** ARISING FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR
+** OTHER DEALINGS IN THE SOFTWARE.
+**
+** For more information, please refer to <http://unlicense.org/>
+*/
 
-package com.kaazing.net.ws.amqp.demo;
+package org.kaazing.net.ws.amqp.demo;
 
 import java.awt.Color;
 import java.awt.EventQueue;
@@ -18,18 +41,18 @@ import java.util.Random;
 import javax.swing.DefaultListModel;
 import javax.swing.border.BevelBorder;
 
-import com.kaazing.net.auth.BasicChallengeHandler;
-import com.kaazing.net.auth.LoginHandler;
-import com.kaazing.net.ws.WebSocketFactory;
-import com.kaazing.net.ws.amqp.AmqpArguments;
-import com.kaazing.net.ws.amqp.AmqpChannel;
-import com.kaazing.net.ws.amqp.AmqpClient;
-import com.kaazing.net.ws.amqp.AmqpProperties;
-import com.kaazing.net.ws.amqp.ChannelAdapter;
-import com.kaazing.net.ws.amqp.ChannelEvent;
-import com.kaazing.net.ws.amqp.ConnectionEvent;
-import com.kaazing.net.ws.amqp.ConnectionListener;
-import com.kaazing.net.ws.amqp.AmqpClientFactory;
+import org.kaazing.net.auth.BasicChallengeHandler;
+import org.kaazing.net.auth.LoginHandler;
+import org.kaazing.net.ws.WebSocketFactory;
+import org.kaazing.net.ws.amqp.AmqpArguments;
+import org.kaazing.net.ws.amqp.AmqpChannel;
+import org.kaazing.net.ws.amqp.AmqpClient;
+import org.kaazing.net.ws.amqp.AmqpProperties;
+import org.kaazing.net.ws.amqp.ChannelAdapter;
+import org.kaazing.net.ws.amqp.ChannelEvent;
+import org.kaazing.net.ws.amqp.ConnectionEvent;
+import org.kaazing.net.ws.amqp.ConnectionListener;
+import org.kaazing.net.ws.amqp.AmqpClientFactory;
 
 public class AmqpPanel extends javax.swing.JPanel implements ActionListener {
